requirements.txt
a2a-sdk==0.3.23
ag-ui-protocol==0.1.11
agent-framework==1.0.0b260123
agent-framework-a2a==1.0.0b260130
agent-framework-ag-ui==1.0.0b260130
agent-framework-anthropic==1.0.0b260130
agent-framework-azure-ai==1.0.0b260123
agent-framework-azure-ai-search==1.0.0b260130
agent-framework-azurefunctions==1.0.0b260130
agent-framework-chatkit==1.0.0b260130
agent-framework-copilotstudio==1.0.0b260130
agent-framework-core==1.0.0b260123
agent-framework-declarative==1.0.0b260130
agent-framework-devui==1.0.0b260130
agent-framework-durabletask==1.0.0b260130
agent-framework-github-copilot==1.0.0b260212
agent-framework-lab==1.0.0b251024
agent-framework-mem0==1.0.0b260130
agent-framework-ollama==1.0.0b260130
agent-framework-orchestrations==1.0.0b260212
agent-framework-purview==1.0.0b260130
agent-framework-redis==1.0.0b260130
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.81.0
anyio==4.12.1
asgiref==3.11.1
asttokens==3.0.1
asyncio==4.0.0
attrs==25.4.0
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext==0.7.5
av==16.1.0
azure-ai-agents==1.2.0b5
azure-ai-inference==1.0.0b9
azure-ai-projects==2.0.0b3
azure-common==1.1.28
azure-core==1.38.1
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.14.6
azure-functions==1.22.1
azure-functions-durable==1.5.0
azure-identity==1.25.2
azure-monitor-opentelemetry==1.8.6
azure-monitor-opentelemetry-exporter==1.0.0b48
azure-search-documents==11.7.0b2
azure-storage-blob==12.28.0
backoff==2.2.1
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudevents==1.12.0
clr_loader==0.2.10
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durabletask==1.3.0
durabletask.azuremanaged==1.3.0
executing==2.2.1
fastapi==0.129.0
frozenlist==1.8.0
furl==2.1.4
github-copilot-sdk==0.1.25
google-api-core==2.29.0
google-auth==2.48.0
google-crc32c==1.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
griffe==1.15.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jsonpath-ng==1.7.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
mem0ai==1.0.4
microsoft-agents-activity==0.7.0
microsoft-agents-copilotstudio-client==0.7.0
microsoft-agents-hosting-core==0.7.0
ml_dtypes==0.5.4
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.1
nest-asyncio==1.6.0
numpy==2.4.2
oauthlib==3.3.1
ollama==0.6.1
openai==2.21.0
openai-agents==0.4.2
openai-chatkit==1.6.1
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.39.0
opentelemetry-instrumentation==0.60b0
opentelemetry-instrumentation-asgi==0.60b0
opentelemetry-instrumentation-dbapi==0.60b0
opentelemetry-instrumentation-django==0.60b0
opentelemetry-instrumentation-fastapi==0.60b0
opentelemetry-instrumentation-flask==0.60b0
opentelemetry-instrumentation-psycopg2==0.60b0
opentelemetry-instrumentation-requests==0.60b0
opentelemetry-instrumentation-urllib==0.60b0
opentelemetry-instrumentation-urllib3==0.60b0
opentelemetry-instrumentation-wsgi==0.60b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b0
orderedmultidict==1.0.2
packaging==26.0
pandas==3.0.1
parse==1.21.0
parso==0.8.6
pathable==0.4.4
pillow==12.1.1
platformdirs==4.9.2
ply==3.11
portalocker==3.2.0
posthog==7.9.1
powerfx==0.0.34
prance==25.4.8.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybars4==0.9.13
pycparser==3.0
pydantic==2.11.10
pydantic-settings==2.13.0
pydantic_core==2.33.2
pyee==13.0.1
Pygments==2.19.2
PyJWT==2.11.0
pylibsrtp==1.0.0
PyMeta3==0.5.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-ulid==3.1.0
pythonnet==3.0.5
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.16.2
redis==7.1.1
redisvl==0.14.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
scipy==1.17.0
semantic-kernel==1.37.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.52.1
tenacity==9.1.4
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.0
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
